Create COVID STRINGS 1 FIRST ROUND.JSON
This is the First Round of Version 1  of the  Covid Strings Contest. The use of a bracket here is especially important because with 8 seeds to start it off there are four matchups in total.  In addition to creating a scoreboard in the background, having a bracket would allow for a player who is not part of a team currently playing to then keep track of other teams and who they might play next should they advance.