Update Covid Strings Second Round.JSON
This Second-Round update is based off the code I sued for the championship game which thankfully was not in the red. I plan on making similar changes for the first round after I put in my game code just to see if part of the problem is a lines issue.