Create SECURITY.md
This security policy is about trying to ensure that there is a way to protect the covid strings experience for gamers as free from hackers or other malicious actors. Given the fact that covid has become somewhat of a divisive subject it is important that we take cyber and physical security seriously by not reposting or spreading covid conspiracy theories or misinformation as truth.